swerve worky

+++ src/main/java/frc/robot/utilities/swerve/SwerveModuleConfig.java
@@ -1,8 +1,8 @@
 package frc.robot.utilities.swerve;
 
-public record SwerveModuleConfig(int driveMotor, boolean driveMotorReversed, int rotationMotor, boolean rotationMotorReversed, int encoder, int encoderOffset, String canbus) {
+public record SwerveModuleConfig(int driveMotor, boolean driveMotorReversed, int rotationMotor, boolean rotationMotorReversed, int encoder, double encoderOffset, String canbus) {
     
-    public SwerveModuleConfig(int driveMotor, boolean driveMotorReversed, int rotationMotor, boolean rotationMotorReversed, int encoder, int encoderOffset){
+    public SwerveModuleConfig(int driveMotor, boolean driveMotorReversed, int rotationMotor, boolean rotationMotorReversed, int encoder, double encoderOffset){
         this(driveMotor, driveMotorReversed, rotationMotor, rotationMotorReversed, encoder, encoderOffset, null);
     }
 }
